pyproject.toml
[project]
name = "spotify-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=3.3.0",
    "kaleido>=1.2.0",
    "matplotlib>=3.10.7",
    "numpy==2.3.3",
    "pandas>=2.3.3",
    "plotly-express>=0.4.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "spotipy>=2.25.1",
    "torch>=2.9.1",
]
